Clean examples
