+++ examples/docker/run.sh
