+++ examples/extractor/external.go
