+++ examples/extractor/external.sh
